% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntree_pkg.R
\docType{package}
\name{dynTree-package}
\alias{dynTree-package}
\alias{_PACKAGE}
\title{dynTree: Landmark tree and ensemble}
\description{
The \code{dynTree} package grow/prune survival trees and ensemble.
}
\section{Introduction}{

The \code{dynTree} package provides implementations to a unified framework for
tree-structured analysis with censored survival outcomes.
}

\section{Methods}{

The package contains functions to construct survival trees and ensemble through
the main function \code{\link{dynTree}}.
}

\seealso{
\code{\link{dynTree}}
}
\author{
\strong{Maintainer}: Sy Han Chiou \email{schiou@utdallas.edu}

Authors:
\itemize{
  \item Yifei Sun \email{ys3072@cumc.columbia.edu}
}

}
