% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dynTree.R
\name{predict.dynTree}
\alias{predict.dynTree}
\title{Predicting based on a \code{dynTree} model.}
\usage{
\method{predict}{dynTree}(object, newdata, control = list(), ...)
}
\arguments{
\item{object}{is an \code{dynTree} object.}

\item{newdata}{is an optional data frame in which to look for variables with which to predict.
If omitted, the fitted predictors are used.
If the covariate observation time is not supplied, covariates will be treated as at baseline.}

\item{control}{a list of control parameters. See 'details' for important special
features of control parameters. See \code{\link{dynTree}} for more details.}

\item{...}{for future developments.}

\item{type}{is an optional character string specifying whether to predict the survival probability or the cumulative hazard rate. (One observation)}
}
\value{
Returns a \code{data.frame} of the predicted survival probabilities or cumulative hazard.
}
\description{
The function gives predicted values with a \code{dynTree} fit.
}
\examples{
data(simDat)
fit <- ltree(Surv(Time, death) ~ z1 + z2, id = id, data = simDat, ensemble = FALSE)

## testing data
newdat <- data.frame(Time = sort(unique(simDat$Time)), 
                     z2 = runif(1))
newdat$z1 <- 1 * (newdat$Time < median(newdat$Time))
head(newdat)

## Predict survival 
pred <- predict(fit, newdat)
plot(pred)

## Predict hazard
pred <- predict(fit, newdat, type = "hazard")
plot(pred)
}
